+++ test/test_matmul_cuda.py
@@ -2,7 +2,6 @@
 
 import contextlib
 import os
-import time
 import unittest
 from itertools import product
 from functools import partial
@@ -21,8 +20,6 @@ from torch.testing import make_tensor
 from torch.testing._internal.common_cuda import (
     blas_library_context,
     PLATFORM_SUPPORTS_BF16,
-    PLATFORM_SUPPORTS_GREEN_CONTEXT,
-    PLATFORM_SUPPORTS_WORKQUEUE_CONFIG,
     SM80OrLater,
     SM90OrLater,
     SM100OrLater,
@@ -1004,126 +1001,6 @@ class TestMatmulCuda(InductorTestCase):
                     op(c, a, mismatch_batch_dim_b, out_dtype=torch.float32)
                 else:
                     op(a, mismatch_batch_dim_b, out_dtype=torch.float32)
-
-
-    @unittest.skipIf(not PLATFORM_SUPPORTS_GREEN_CONTEXT, "Green contexts are not supported")
-    @serialTest()
-    def test_greencontext_carveout(self):
-        a = torch.randn(4096, 4096, device='cuda', dtype=torch.bfloat16)
-        ctx = torch.cuda.green_contexts.GreenContext.create(num_sms=1, device_id=0)
-        ctx.set_context()
-        torch.matmul(a, a)
-        torch.cuda.synchronize()
-        t0 = time.perf_counter()
-        partial_res = torch.matmul(a, a)
-        torch.cuda.synchronize()
-        t1 = time.perf_counter()
-        ctx.pop_context()
-        torch.matmul(a, a)
-        torch.cuda.synchronize()
-        t2 = time.perf_counter()
-        full_res = torch.matmul(a, a)
-        torch.cuda.synchronize()
-        t3 = time.perf_counter()
-        self.assertEqual(partial_res, full_res)
-        self.assertGreater(t1 - t0, t3 - t2)
-
-    @unittest.skipIf(not PLATFORM_SUPPORTS_GREEN_CONTEXT, "Green contexts are not supported")
-    @serialTest()
-    def test_greencontext_stream_carveout(self):
-        a = torch.randn(4096, 4096, device='cuda', dtype=torch.bfloat16)
-        ctx = torch.cuda.green_contexts.GreenContext.create(num_sms=1, device_id=0)
-        ctx_stream = ctx.Stream()
-        with torch.cuda.stream(ctx_stream):
-            torch.matmul(a, a)
-            torch.cuda.synchronize()
-            t0 = time.perf_counter()
-            partial_res = torch.matmul(a, a)
-            torch.cuda.synchronize()
-            t1 = time.perf_counter()
-        torch.matmul(a, a)
-        torch.cuda.synchronize()
-        t2 = time.perf_counter()
-        full_res = torch.matmul(a, a)
-        torch.cuda.synchronize()
-        t3 = time.perf_counter()
-        self.assertEqual(partial_res, full_res)
-        self.assertGreater(t1 - t0, t3 - t2)
-
-    @unittest.skipIf(not PLATFORM_SUPPORTS_GREEN_CONTEXT, "Green contexts are not supported")
-    @serialTest()
-    def test_greencontext_graphs(self):
-        a = torch.randn(4096, 4096, device='cuda', dtype=torch.bfloat16)
-        ctx = torch.cuda.green_contexts.GreenContext.create(num_sms=1, device_id=0)
-        ctx.set_context()
-        partial_res = torch.matmul(a, a)
-        ctx.pop_context()
-        full_res = torch.matmul(a, a)
-        full_res.zero_()
-        partial_res.zero_()
-        torch.cuda.synchronize()
-
-        g = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g):
-            ctx.set_context()
-            partial_res = torch.matmul(a, a)
-            ctx.pop_context()
-            full_res = torch.matmul(a, a)
-        g.replay()
-        self.assertEqual(partial_res, full_res)
-
-    @unittest.skipIf(not PLATFORM_SUPPORTS_WORKQUEUE_CONFIG, "Workqueue config is not supported")
-    @serialTest()
-    def test_greencontext_workqueue_concurrency_limit(self):
-        # Note: this test simply tests green context workqueue concurrency limiting,
-        # It doesn't test matmuls specifically, but is kept here along with other
-        # green context functional tests.
-        n_streams = 4
-        GreenContext = torch.cuda.green_contexts.GreenContext
-        max_wq = GreenContext.max_workqueue_concurrency(device_id=0)
-        if max_wq < n_streams:
-            self.skipTest(f"Device has {max_wq} workqueue(s), need >{n_streams} to test concurrency limiting")
-
-        wq_events = [torch.cuda.Event(enable_timing=True) for _ in range(n_streams * 2)]
-
-        def run_multi_stream_sleep(streams):
-            for i, s in enumerate(streams):
-                with torch.cuda.stream(s):
-                    # note we need to record timing events to ensure that the
-                    # workqueue is actually used
-                    wq_events[i * 2].record(s)
-                    # 100M cycles is enough on all currently supported GPUs
-                    # to ensure that the test runs significantly longer than
-                    # host overhead to 1) activate the different streams,
-                    # 2) record timing events, and 3) synchronize with the GPU.
-                    torch.cuda._sleep(100_000_000)
-                    wq_events[i * 2 + 1].record(s)
-            torch.cuda.synchronize()
-
-        # Baseline: n_streams streams from default context, full workqueue concurrency
-        baseline_streams = [torch.cuda.Stream() for _ in range(n_streams)]
-        # note: torch.cuda.synchronize() will wait for all kernels on all streams
-        # so we can safely use CPU based timing here.
-        t0 = time.perf_counter()
-        run_multi_stream_sleep(baseline_streams)
-        t1 = time.perf_counter()
-        baseline_time = t1 - t0
-
-        # Green context with workqueue concurrency limited to 1
-        ctx = GreenContext.create(
-            workqueue_scope="balanced",
-            workqueue_concurrency_limit=1,
-            device_id=0,
-        )
-        ctx.set_context()
-        green_streams = [ctx.Stream() for _ in range(n_streams)]
-        t2 = time.perf_counter()
-        run_multi_stream_sleep(green_streams)
-        t3 = time.perf_counter()
-        ctx.pop_context()
-        limited_time = t3 - t2
-
-        self.assertGreater(limited_time, baseline_time)
 
 
 @unittest.skipIf(TEST_WITH_ROCM, "ROCm doesn't support CUTLASS")
